feat: add mathjax.
The mathjax library apparently cannot be simply added via something like
'pnpm add mathjax'. The tutorials often force into downloading via CDN,
which is a no-go, since the app wouldn't be usable without internet.

The solution is to copy the sources from the mathjax repository (es5
folder) and copy them into the app itself. At the very least we should
automatize this somehow, and not copy it manually. Also, it is probable
that the mathjax has some files that we don't need.