feat: add dialog for synchronization + cleanup.
Adds dialog for triggering synchronization on a folder. The folders are
not watched and the synchronization has to be run every time by hand.
This is WIP and will be reworked in the future.

The markdown parsing was copied from the old version of the repository
and was rewritten using tokio::BufReader to be async.

Also, some code/import cleanup has been performed.