feat: Add sync to frontend, make errors more clear, add nav.
A new 'settings' page added to frontend, for now only used for loading
and syncing cards from files. The error messages on rust part were made
more clear so that user understands what is going on. A toast library
was added for frontend.

A simple, bare bones navigation bar was added so that one can navigate
through the app.