feat: add more customization to settings page, prep. for more features.
Add a list of folders that can be synchronized, and a button that
triggers synchronization accross all of them. A checkbox to switch
between randomized order of cards when reviewing.

Two more features were added that are currently not implemented, but the
FE part is prepared. A selection of repetition algorithm and a checkbox
to trigger synchronization on app startup.